analytic_service/requirements.txt
fastapi
uvicorn
clickhouse-connect
pyyaml
pydantic